requirements.txt
streamlit
azure-search-documents
openai
dotenv